refactor: Extract modal component
