--- public/donationsApi.json
@@ -1,157 +1,0 @@
-{
-  "list": [
-    {
-      "id": 580,
-      "idolId": 3708,
-      "title": "도훈 생일 축하",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 2111,
-      "createdAt": "2024-09-26T16:43:23.180Z",
-      "deadline": "2024-09-28T00:00:00.000Z",
-      "status": true,
-      "idol": {
-        "id": 3708,
-        "name": "도훈",
-        "gender": "male",
-        "group": "투어스",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727365747712/dohoon.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 579,
-      "idolId": 3700,
-      "title": "엔시티드림 마크 생일 축하",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 10000,
-      "createdAt": "2024-09-26T16:33:50.958Z",
-      "deadline": "2024-09-29T23:59:59.000Z",
-      "status": false,
-      "idol": {
-        "id": 3700,
-        "name": "마크",
-        "gender": "male",
-        "group": "엔시티드림",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727362424463/MARK.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 578,
-      "idolId": 3693,
-      "title": "라이즈 쇼타로 생일 축하",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:33:14.834Z",
-      "deadline": "2024-09-28T16:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3693,
-        "name": "쇼타로",
-        "gender": "male",
-        "group": "라이즈",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727361690348/shotaro.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 577,
-      "idolId": 3691,
-      "title": "있지 채령 지하철 광고",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:30:59.917Z",
-      "deadline": "2025-09-29T23:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3691,
-        "name": "채령",
-        "gender": "female",
-        "group": "있지",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727361201458/chaeryeong.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 576,
-      "idolId": 3668,
-      "title": "블랙핑크 제니 지하철 광고",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:27:20.629Z",
-      "deadline": "2025-09-29T23:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3668,
-        "name": "제니",
-        "gender": "female",
-        "group": "블랙핑크",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727349152987/jenni.jfif",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 575,
-      "idolId": 3678,
-      "title": "르세라핌 채원 지하철 광고",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:24:28.192Z",
-      "deadline": "2025-09-29T23:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3678,
-        "name": "김채원",
-        "gender": "female",
-        "group": "르세라핌",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727349712588/kimchaewon.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 574,
-      "idolId": 3673,
-      "title": "원영이 20번째 생일",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:23:10.219Z",
-      "deadline": "2025-09-29T23:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3673,
-        "name": "장원영",
-        "gender": "female",
-        "group": "아이브",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727349508092/JANGWONYOUNG.webp",
-        "totalVotes": 0
-      }
-    },
-    {
-      "id": 573,
-      "idolId": 3682,
-      "title": "뉴진스 민지 지하철 광고",
-      "subtitle": "강남역 광고",
-      "targetDonation": 6000,
-      "receivedDonations": 0,
-      "createdAt": "2024-09-26T16:21:22.411Z",
-      "deadline": "2025-09-29T23:59:59.000Z",
-      "status": true,
-      "idol": {
-        "id": 3682,
-        "name": "민지",
-        "gender": "female",
-        "group": "뉴진스",
-        "profilePicture": "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/Fandom-K/idol/1727349843187/minji.webp",
-        "totalVotes": 0
-      }
-    }
-  ],
-  "nextCursor": null
-}
