+++ src/hooks/useModalPrev.jsx
@@ -1,21 +1,17 @@
 import { useState } from 'react';
 
-export default function useModal() {
+export function useModal(Modal) {
   const [isOpen, setIsOpen] = useState(false);
 
   const openModal = () => {
     setIsOpen(true);
-    document.body.style.overflow = 'hidden';
   };
 
   const closeModal = () => {
     setIsOpen(false);
-    document.body.style.overflow = 'unset';
   };
 
-  return {
-    isOpen,
-    openModal,
-    closeModal,
-  };
+  const wrappedModal = () => isOpen && <Modal onClose={closeModal}></Modal>;
+
+  return { openModal, Modal: wrappedModal };
 }
