+++ src/pages/listPage/Donation/components/DonationCard.jsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import Button from './Button';
 import styles from './DonationCard.module.css';
 import ProgressBar from './ProgressBar';
-import Modal from './Modal';
+import Modal from '../../../../components/Modal';
 import DonationModalContent from './DonationModalContent';
-import useModal from '../hooks/useModal';
+import useModal from '../../../../hooks/useModal';
 
 export default function DonationCard({
   id,
